requirements.txt
zeep
pydantic
python-dotenv
google-generativeai
pandas
openpyxl
requests
PyQt6
pyinstaller
